21.10.23: some minor fixes to commands and commons

+++ Functions.cs
@@ -80,8 +80,14 @@ namespace Entropy
                 {
                     try
                     {
+                        if (process.ProcessName.ToString() == "Entropy")
+                        {
+                            Utilities.EntropyWrite(ConsoleColor.Red, $"What????");
+                            return;
+                        }
                         if (process.Id.ToString() == argument)
                         {
+
                             process.Kill();
                             Utilities.EntropyWrite(ConsoleColor.Green, $"{process.Id}:::{process.ProcessName} was succesfully terminated.");
                             isFinded = true;
